Expand laravel/nova version requirements
If you've installed Nova into your app using the non-composer based instructions (which is common for teams working on a codebase) then the existing laravel/nova 1.0 constraint isn't satisfied. By re-writing as '*' the package can be added. 

I will grant, this is not an elegant solution. If there is a better solution I am anxious to know it.